organized folders/file structure and updated cooresponding routing in App.jsx and main.jsx. Added Footer Folder/Footer.jsx and style.css. Also added global background color for website based on Figma
