+++ client/src/App.css
@@ -2,5 +2,6 @@
   font-family: 'Arial', sans-serif;
     margin: 0;
     padding: 0;
+    background:#41294A; 
 }
